Avoid overflow in CUDA range batch count

+++ core/src/Cuda/Kokkos_Cuda_Parallel_Range.hpp
@@ -8,6 +8,7 @@
 #if defined(KOKKOS_ENABLE_CUDA)
 
 #include <algorithm>
+#include <cstdint>
 #include <string>
 
 #include <Kokkos_Parallel.hpp>
@@ -91,7 +92,7 @@ class ParallelFor<FunctorType, Kokkos::RangePolicy<Traits...>, Kokkos::Cuda> {
     if constexpr (StaticBatchSize::batch_size != 1) {
       constexpr typename Policy::index_type batch_size =
           StaticBatchSize::batch_size;
-      nwork = (range + batch_size - 1) / batch_size;
+      nwork = (uint64_t(range) + batch_size - 1) / batch_size;
     }
     cudaFuncAttributes attr =
         CudaParallelLaunch<ParallelFor, LaunchBounds>::get_cuda_func_attributes(
